Fix norm errors in file/

+++ file/read.c
@@ -24,7 +24,7 @@ int		read_file(const t_hs filename, t_glist_hs *lines)
 {
   char		*data;
   int		fd;
-  size_t	file_length;
+  ssize_t	file_length;
   int		i;
   t_hs		ths;
   t_glist_hs	new_lines;
@@ -43,7 +43,7 @@ int		read_file(const t_hs filename, t_glist_hs *lines)
     }
   close_file(fd);
   ths = hs(data);
-  if (hs_length(ths) != file_length)
+  if ((int)hs_length(ths) != file_length)
     return (-1);
   new_lines = hs_split(ths, hs("\n"));
   glist_hs_append_all(lines, &new_lines);
